move JdbcUtils to hdata-jdbc module
